pocs/beeai-showcase-security/requirements.txt
beeai-framework>=0.1.79
rich>=13.7.0
python-dotenv>=1.0.0
openai>=1.0.0
pydantic>=2.0.0
